// src/@coglite/appstore/listing/types/IListingReview.ts
import { IUserProfile } from "../../user/types/IUserProfile";

interface IListingReview {
    id?: number;
    author?: IUserProfile;
    listing?: number;
    rate?: number;
    text?: string;
    edited_date?: string;
    created_date?: string;
    review_parent?: number;
    review_response?: IListingReview[];
}

export { IListingReview }